\name{clusterings}
\alias{clusterings}
\docType{data}
\title{clusterings}
\description{This list defines a certain number of clustering algorithms which
we can compare. Each element is a list with 2 elements: [[1]] is
the clustering method name, [[2]] is a function which takes a data
matrix m and number of clusters k and then returns a vector of
integer cluster assignments.}
\usage{"clusterings"}





