\name{clusterpath-package}
\alias{clusterpath-package}
\docType{package}
\title{Fast agglomerative clustering using convex fusion penalties}
\description{Implements the clusterpath, a clustering method in which all
the clusters are solutions to a convex optimization problem with fusion
penalties. This package is a wrapper around the C++ implementation.}
\details{
\tabular{ll}{Package: \tab clusterpath\cr
Type: \tab Package\cr
Title: \tab Fast agglomerative clustering using convex fusion penalties\cr
Version: \tab 1.0\cr
Date: \tab 2011-05-10\cr
Author: \tab Toby Dylan Hocking\cr
Maintainer: \tab Toby Dylan Hocking <toby.hocking@inria.fr>\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
Depends: \tab Rcpp (>= 0.8.0), lattice, ggplot2, plyr, reshape, latticeExtra, Matrix\cr
Suggests: \tab multicore, rgl, animation, directlabels, mclust\cr
LinkingTo: \tab Rcpp\cr
SystemRequirements: \tab GNU make\cr}
}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}

\keyword{ package }


