\name{cut.clusterpath.labels}
\alias{cut.clusterpath.labels}
\title{cut clusterpath labels}
\description{given a clusterpath tree, cut it and give the labels}
\usage{
\method{cut}{clusterpath.labels}(x, k, ...)
}
\arguments{
  \item{x}{
}
  \item{k}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking <toby.hocking@inria.fr>}





