\name{cut.clusterpath}
\alias{cut.clusterpath}
\title{cut clusterpath}
\description{Given a clusterpath data frame, cut it at a certain number of
clusters and return the tree up to that point}
\usage{
\method{cut}{clusterpath}(x, k, ...)
}
\arguments{
  \item{x}{
}
  \item{k}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking <toby.hocking@inria.fr>}





