\name{cvxmod.available}
\alias{cvxmod.available}
\title{cvxmod available}
\description{Test if cvxmod is working on this system.}
\usage{cvxmod.available(python = python.command())}
\arguments{
  \item{python}{How should python be invoked?}
}

\value{TRUE if cvxmod is available, FALSE otherwise.}

\author{Toby Dylan Hocking}





