\name{e.exponential.auto}
\alias{e.exponential.auto}
\title{e exponential auto}
\description{adaptive eigenvector scaling function that takes the biggest
bandwidth that gives you more than 1 nonzero vector.}
\usage{e.exponential.auto(eigenvalues)}
\arguments{
  \item{eigenvalues}{
}
}



\author{Toby Dylan Hocking}





