\name{e.exponential}
\alias{e.exponential}
\title{e exponential}
\description{eigenvector scaling function with fixed bandwidth.}
\usage{
e.exponential(evals, bw = 1e+09)
}
\arguments{
  \item{evals}{
}
  \item{bw}{
}
}



\author{Toby Dylan Hocking <toby.hocking@inria.fr>}





