\name{exponential.eigenvectors}
\alias{exponential.eigenvectors}
\title{exponential eigenvectors}
\description{scale the eigenvectors using eigenvalues and the exponential
function}
\usage{exponential.eigenvectors(nonzero.vectors, nonzero.values, bandwidth = 1)}
\arguments{
  \item{nonzero.vectors}{
}
  \item{nonzero.values}{
}
  \item{bandwidth}{
}
}



\author{Toby Dylan Hocking}





