\name{gendata}
\alias{gendata}
\title{gendata}
\description{generate a multivariate normal mixture}
\usage{
gendata(N = 5, K = 2, D = 3, SD = 0.1)
}
\arguments{
  \item{N}{
}
  \item{K}{
}
  \item{D}{
}
  \item{SD}{
}
}



\author{Toby Dylan Hocking <toby.hocking@inria.fr>}





