\name{get.old.best}
\alias{get.old.best}
\title{get old best}
\description{try a bunch of gamma values in spectral clustering and get the one
that gives the best reconstruction error.}
\usage{
get.old.best(m, k, evec.transform)
}
\arguments{
  \item{m}{
}
  \item{k}{
}
  \item{evec.transform}{
}
}



\author{Toby Dylan Hocking <toby.hocking@inria.fr>}





