\name{norm.rand.index}
\alias{norm.rand.index}
\title{norm rand index}
\description{measure of correspondence between partitions (Hubert and Arabie
1985) 1=perfect, 0=completely random, or just the same label for
everybody.}
\usage{norm.rand.index(klass, guess)}
\arguments{
  \item{klass}{
}
  \item{guess}{
}
}



\author{Toby Dylan Hocking}





