\name{clusterpath-package}
\alias{clusterpath-package}
\alias{clusterpath}
\docType{package}
\title{Fast agglomerative convex clustering, non-Rcpp implementation}

\description{Implements the clusterpath, a clustering method in which all
the clusters are solutions to a convex optimization problem with fusion
penalties.
This package uses standard R/C headers to wrap the C++ implementation.}

\details{
The DESCRIPTION file:
\packageDESCRIPTION{clusterpath}
\packageIndices{clusterpath}
}
\author{Toby Dylan Hocking}

\keyword{ package }


