\name{clusterpath-package}
\alias{clusterpath-package}
\docType{package}
\title{Fast agglomerative convex clustering, non-Rcpp implementation}
\description{Implements the clusterpath, a clustering method in which all
the clusters are solutions to a convex optimization problem with fusion
penalties.
This package uses standard R/C headers to wrap the C++ implementation.}
\details{
\tabular{ll}{Package: \tab clusterpath\cr
Type: \tab Package\cr
Title: \tab Fast agglomerative convex clustering, non-Rcpp implementation\cr
Version: \tab 1.1\cr
Date: \tab 2013-01-24\cr
Author: \tab Toby Dylan Hocking\cr
Maintainer: \tab Toby Dylan Hocking <toby.hocking@inria.fr>\cr
License: \tab GPL-3\cr
Depends: \tab lattice, ggplot2, plyr, reshape2,
latticeExtra, Matrix, grid\cr
Suggests: \tab rgl, directlabels, mclust, multicore\cr}
}
\author{Toby Dylan Hocking}

\keyword{ package }


